Move version/version.go to api module
