+++ packages/kn-plugin-workflow/go.mod
@@ -30,7 +30,6 @@ require (
 
 require (
 	github.com/Microsoft/go-winio v0.6.1 // indirect
-	github.com/apache/incubator-kie-tools/packages/sonataflow-operator v0.0.0-20250508125535-082fda209c4b // indirect
 	github.com/beorn7/perks v1.0.1 // indirect
 	github.com/cespare/xxhash/v2 v2.3.0 // indirect
 	github.com/containerd/log v0.1.0 // indirect
