+++ packages/sonataflow-operator/internal/controller/platform/k8s.go
@@ -23,10 +23,11 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/apache/incubator-kie-tools/packages/sonataflow-operator/api/metadata"
-	"github.com/apache/incubator-kie-tools/packages/sonataflow-operator/version"
+	"github.com/apache/incubator-kie-tools/packages/sonataflow-operator/api/version"
 
 	"k8s.io/klog/v2"
+
+	"github.com/apache/incubator-kie-tools/packages/sonataflow-operator/api/metadata"
 
 	"github.com/imdario/mergo"
 	appsv1 "k8s.io/api/apps/v1"
