+++ packages/sonataflow-operator/internal/controller/profiles/dev/object_creators_dev_test.go
@@ -22,7 +22,7 @@ package dev
 import (
 	"testing"
 
-	"github.com/apache/incubator-kie-tools/packages/sonataflow-operator/version"
+	"github.com/apache/incubator-kie-tools/packages/sonataflow-operator/api/version"
 
 	"github.com/stretchr/testify/assert"
 	v1 "k8s.io/api/core/v1"
