+++ packages/sonataflow-operator/workflowproj/operator.go
@@ -22,7 +22,7 @@ package workflowproj
 import (
 	"fmt"
 
-	"github.com/apache/incubator-kie-tools/packages/sonataflow-operator/version"
+	"github.com/apache/incubator-kie-tools/packages/sonataflow-operator/api/version"
 
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
